requirements.txt
Flask
tinydb
request